require srt callbacks be `Send` and `Sync` (#175)
The `Send` reflects the reality that the SRT code internally passes
these to its listener thread and runs there. Using a non-`Send` callback
would be unsound!

I also added the `Sync` bound to allow the `AsyncListener` to be `Sync`.
I believe this isn't strictly necessary: conceptually the
ownership is handed to the SRT listener thread, then back later to
run the destructor, so we could restructure `_callback`'s type a bit to
be more permissive. But this is an easy way to solve the problem and
seems harmless.