do not set `vbv_buffer_size` (#212)
With Quadra firmware 4.9.2, setting this parameter no longer appears
to be necessary, and in fact is harmful. Setting it to 0 as this code
was doing for larger bitrates appears to effectively negate the bitrate
target, with a significant difference in quality. The `ffmpeg` command
below produced an encoding with 18 Mbps (compare to target 200 Mbps) and
a VMAF (harmonic mean) of 79.4. By removing the `:vbvBufferSize=0`,
we instead reached 185 Mbps and 86.5, respectively.

```
ffmpeg -y -nostdin -hide_banner \
    -video_size 7680x4320 -r 30 -pixel_format yuv420p -i avs8kraw30.yuv420p.yuv \
    -c:v h265_ni_quadra_enc -xcoder-params 'intraPeriod=90:RcEnable=1:bitrate=200000000:vbvBufferSize=0:gopPresetIdx=9:multicoreJointMode=1' avs8kraw30.cur.mp4
```